We don't seem to have tests for --js-library in core suite, so move other.test_js_libraries to the core test suite so that it gets tested in all modes.
